make state-handling methods not abstract in AbsstracInvokable
