+++ flink-tests/src/test/scala/org/apache/flink/api/scala/runtime/jobmanager/JobManagerLeaderSessionIDITSuite.scala
@@ -23,9 +23,7 @@ import java.util.UUID
 import akka.actor.ActorSystem
 import akka.actor.Status.Success
 import akka.testkit.{ImplicitSender, TestKit}
-import org.apache.flink.api.common.ExecutionConfig
 import org.apache.flink.runtime.akka.{AkkaUtils, ListeningBehaviour}
-import org.apache.flink.runtime.checkpoint.{CheckpointMetaData, CheckpointMetrics, CheckpointOptions}
 import org.apache.flink.runtime.execution.Environment
 import org.apache.flink.runtime.jobgraph.tasks.AbstractInvokable
 import org.apache.flink.runtime.jobgraph.{JobGraph, JobVertex}
@@ -98,29 +96,6 @@ class BlockingUntilSignalNoOpInvokable(environment: Environment, taskStateHandle
     BlockingUntilSignalNoOpInvokable.lock.synchronized{
       BlockingUntilSignalNoOpInvokable.lock.wait()
     }
-  }
-
-  override def triggerCheckpoint(checkpointMetaData: CheckpointMetaData,
-                                  checkpointOptions: CheckpointOptions): Boolean = {
-    throw new UnsupportedOperationException(
-      String.format("triggerCheckpoint not supported by %s", this.getClass.getName))
-  }
-
-  override def triggerCheckpointOnBarrier(checkpointMetaData: CheckpointMetaData,
-                                          checkpointOptions: CheckpointOptions,
-                                          checkpointMetrics: CheckpointMetrics): Unit = {
-    throw new UnsupportedOperationException(
-      String.format("triggerCheckpointOnBarrier not supported by %s", this.getClass.getName))
-  }
-
-  override def abortCheckpointOnBarrier(checkpointId: Long, cause: Throwable): Unit = {
-    throw new UnsupportedOperationException(
-      String.format("abortCheckpointOnBarrier not supported by %s", this.getClass.getName))
-  }
-
-  override def notifyCheckpointComplete(checkpointId: Long): Unit = {
-    throw new UnsupportedOperationException(
-      String.format("notifyCheckpointComplete not supported by %s", this.getClass.getName))
   }
 }
